Move the root reducer to the modules folder
