+++ src/modules/reducer.js
@@ -2,12 +2,12 @@
 
 import { combineReducers } from 'redux'
 
-import { core } from '../modules/Core/reducer.js'
-import { exchangeRates } from '../modules/ExchangeRates/reducer.js'
-import { ui } from '../modules/UI/reducer.js'
 import { contactsReducer as contacts } from '../reducers/contacts/indexContacts.js'
-import cryptoExchange from '../reducers/CryptoExchangeReducer'
+import cryptoExchange from '../reducers/CryptoExchangeReducer.js'
 import { permissionsReducer as permissions } from '../reducers/permissions/indexPermissions.js'
+import { core } from './Core/reducer.js'
+import { exchangeRates } from './ExchangeRates/reducer.js'
+import { ui } from './UI/reducer.js'
 
 export { core, ui, cryptoExchange, exchangeRates, permissions, contacts }
